+++ presentation/src/main/java/com/shypolarbear/presentation/ui/feed/FeedFragment.kt
@@ -4,44 +4,19 @@ import androidx.fragment.app.viewModels
 import com.shypolarbear.presentation.R
 import com.shypolarbear.presentation.base.BaseFragment
 import com.shypolarbear.presentation.databinding.FragmentFeedBinding
-import com.shypolarbear.presentation.ui.feed.adapter.FeedPostAdapter
-import com.shypolarbear.presentation.util.FunctionUtil
-import com.skydoves.powermenu.PowerMenuItem
+import com.shypolarbear.presentation.ui.feed.feedDetail.FeedDetailFragment
 
-class FeedFragment: BaseFragment<FragmentFeedBinding, FeedViewModel> (
+class FeedFragment: BaseFragment<FragmentFeedBinding, FeedViewModel>(
     R.layout.fragment_feed
-){
-
-    companion object {
-        const val POWER_MENU_OFFSET_X = -290
-        const val POWER_MENU_OFFSET_Y = 0
-    }
+) {
 
     override val viewModel: FeedViewModel by viewModels()
-    private val feedSortItems: List<PowerMenuItem> by lazy {
-        listOf(
-            PowerMenuItem(requireContext().getString(R.string.feed_post_property_recent)),
-            PowerMenuItem(requireContext().getString(R.string.feed_post_property_recent_best)),
-            PowerMenuItem(requireContext().getString(R.string.feed_post_property_best))
-        )
-    }
 
     override fun initView() {
-        val functionUtil = FunctionUtil(binding.root.context, feedSortItems, viewLifecycleOwner )
-
-        binding.ivFeedToolbarSort.setOnClickListener {
-            functionUtil.setMenu(binding.ivFeedToolbarSort)
-        }
-
-        viewModel.loadFeedPost()
-        setFeedPost()
+        childFragmentManager.beginTransaction()
+            .replace(R.id.feed, FeedDetailFragment())
+            .addToBackStack(null)
+            .commit()
     }
 
-    private fun setFeedPost() {
-        val feedPostAdapter = FeedPostAdapter(viewLifecycleOwner)
-        binding.rvFeedPost.adapter = feedPostAdapter
-        viewModel.feedPost.observe(viewLifecycleOwner) {
-            feedPostAdapter.submitList(it)
-        }
-    }
 }
