+++ presentation/src/main/java/com/shypolarbear/presentation/ui/feed/adapter/FeedPostAdapter.kt
@@ -2,6 +2,7 @@ package com.shypolarbear.presentation.ui.feed.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.LifecycleOwner
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
@@ -9,13 +10,16 @@ import com.shypolarbear.domain.model.feed.FeedPost
 import com.shypolarbear.presentation.databinding.ItemFeedBinding
 import com.shypolarbear.presentation.ui.feed.viewholder.FeedPostViewHolder
 
-class FeedPostAdapter(private val viewLifeCycleOwner: LifecycleOwner): ListAdapter<FeedPost, FeedPostViewHolder>(FeedPostDiffCallback()) {
+class FeedPostAdapter(
+    private val viewLifeCycleOwner: LifecycleOwner,
+    private val fragmentManager: FragmentManager
+    ): ListAdapter<FeedPost, FeedPostViewHolder>(FeedPostDiffCallback()) {
 
     private lateinit var binding : ItemFeedBinding
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FeedPostViewHolder {
         binding = ItemFeedBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return FeedPostViewHolder(binding, viewLifeCycleOwner)
+        return FeedPostViewHolder(binding, viewLifeCycleOwner, fragmentManager)
     }
 
     override fun onBindViewHolder(holder: FeedPostViewHolder, position: Int) {
