+++ presentation/src/main/res/layout/activity_main.xml
@@ -12,7 +12,7 @@
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fragment1"
-            android:name="com.shypolarbear.presentation.ui.feed.FeedFragment"
+            android:name="com.shypolarbear.presentation.ui.feed.feedTotal.FeedTotalFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             app:layout_constraintBottom_toTopOf="@+id/bottom_navigation_bar"
